requirements.txt
rembg
pillow
numpy
opencv-python
